refactor: #90 앰프 package 분리
