+++ src/main/java/com/ajou/hertz/domain/instrument/amplifier/constant/AmplifierType.java
@@ -1,0 +1,7 @@
+package com.ajou.hertz.domain.instrument.amplifier.constant;
+
+public enum AmplifierType {
+
+	GUITAR,
+	BASS
+}
