+++ src/main/java/com/ajou/hertz/domain/instrument/amplifier/constant/AmplifierUsage.java
@@ -1,4 +1,4 @@
-package com.ajou.hertz.domain.instrument.constant;
+package com.ajou.hertz.domain.instrument.amplifier.constant;
 
 import lombok.AllArgsConstructor;
