+++ src/main/java/com/ajou/hertz/domain/instrument/amplifier/dto/response/AmplifierResponse.java
@@ -1,14 +1,16 @@
-package com.ajou.hertz.domain.instrument.dto.response;
+package com.ajou.hertz.domain.instrument.amplifier.dto.response;
 
 import java.util.List;
 
 import com.ajou.hertz.common.dto.response.AddressResponse;
-import com.ajou.hertz.domain.instrument.constant.AmplifierBrand;
-import com.ajou.hertz.domain.instrument.constant.AmplifierType;
-import com.ajou.hertz.domain.instrument.constant.AmplifierUsage;
+import com.ajou.hertz.domain.instrument.amplifier.constant.AmplifierBrand;
+import com.ajou.hertz.domain.instrument.amplifier.constant.AmplifierType;
+import com.ajou.hertz.domain.instrument.amplifier.constant.AmplifierUsage;
 import com.ajou.hertz.domain.instrument.constant.InstrumentCategory;
 import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
-import com.ajou.hertz.domain.instrument.dto.AmplifierDto;
+import com.ajou.hertz.domain.instrument.amplifier.dto.AmplifierDto;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentImageResponse;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentResponse;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
