+++ src/main/java/com/ajou/hertz/domain/instrument/amplifier/strategy/AmplifierCreationStrategy.java
@@ -1,7 +1,8 @@
-package com.ajou.hertz.domain.instrument.strategy;
+package com.ajou.hertz.domain.instrument.amplifier.strategy;
 
-import com.ajou.hertz.domain.instrument.dto.request.CreateNewAmplifierRequest;
-import com.ajou.hertz.domain.instrument.entity.Amplifier;
+import com.ajou.hertz.domain.instrument.amplifier.dto.request.CreateNewAmplifierRequest;
+import com.ajou.hertz.domain.instrument.amplifier.entity.Amplifier;
+import com.ajou.hertz.domain.instrument.strategy.InstrumentCreationStrategy;
 import com.ajou.hertz.domain.user.entity.User;
 
 public class AmplifierCreationStrategy implements InstrumentCreationStrategy<Amplifier, CreateNewAmplifierRequest> {
