--- src/main/java/com/ajou/hertz/domain/instrument/constant/AmplifierType.java
@@ -1,7 +1,0 @@
-package com.ajou.hertz.domain.instrument.constant;
-
-public enum AmplifierType {
-
-	GUITAR,
-	BASS
-}
